// src/app/model/image-batch.ts
export class ImageBatch {
    id: number;
    date: Date;
    individualId: number;
    individualLabel: string;
    individualFullyQualifiedLabel: string;
    experimentalGroupLabel: string;
    protocolLabel: string;
    protocolTitle: string;
    fileCount: number;
    applyStitching: boolean;
    userEmail: string;
    userFirstName: string;
    userLastName: string;
}